+++ scilab/modules/scicos/sci_gateway/cpp/sci_scicos_log.cpp
@@ -1,5 +1,5 @@
 /*
- *  Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
+ *  Scilab ( https://www.scilab.org/ ) - This file is part of Scilab
  *  Copyright (C) 2014-2014 - Scilab Enterprises - Clement DAVID
  *
  * Copyright (C) 2012 - 2016 - Scilab Enterprises
@@ -101,7 +101,7 @@ types::Function::ReturnValue sci_scicos_log(types::typed_list &in, int _iRetCoun
 
         types::Double* refCounts = new types::Double((int) objects.size(), 3);
 
-        for (int i = 0; i < objects.size(); ++i)
+        for (int i = 0; i < (int) objects.size(); ++i)
         {
             model::BaseObject* o = controller.getBaseObject(objects[i]);
